updated reqs, moved things around
