+++ .travis.yml
@@ -1,7 +1,7 @@
 language: python
 python:
-  - "3.4"
-  - "3.5"
+  - "3.6"
+  - "3.7"
 branches:
   only:
     - master
