+++ opsy/app.py
@@ -1,7 +1,5 @@
 import logging
 from logging.handlers import WatchedFileHandler
-from logging import StreamHandler
-import sys
 from flask import Flask
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from opsy.main import core_main
@@ -9,7 +7,7 @@ from opsy.api import core_api
 from opsy.flask_extensions import configure_extensions
 from opsy.utils import OpsyJSONEncoder, load_plugins
 from opsy.config import load_config, validate_config
-from opsy.auth.access import needs
+from opsy.access import needs
 
 
 def create_app(config_file):
@@ -47,15 +45,16 @@ def create_app(config_file):
 
 def create_logging(app):
     if not app.debug:
-        log_handlers = [StreamHandler(sys.stdout)]
+        log_handlers = []
         log_file = app.config.opsy['log_file']
         if log_file:
             log_handlers.append(WatchedFileHandler(log_file))
         formatter = logging.Formatter(
             "%(asctime)s %(process)d %(levelname)s %(module)s - %(message)s")
         for log_handler in log_handlers:
-            log_handler.setFormatter(formatter)
             app.logger.addHandler(log_handler)
+        for log_handler in app.logger.handlers:
+            log_handler.setFormatter(formatter)
         app.logger.setLevel(logging.INFO)
