+++ requirements.txt
@@ -1,17 +1,19 @@
+aiohttp==3.5.1
+apscheduler==3.5.3
 Flask==0.12
-aiohttp==0.21.6
-humanize==0.5.1
 Flask-INIConfig==0.1.0
-Flask-SQLAlchemy==2.1
-Flask-JSGlue==0.3
-Flask-RESTful==0.3.5
-Flask-Login==0.4.0
+Flask-JSGlue==0.3.1
+flask-ldap3-login==0.9.16
+Flask-Login==0.4.1
+flask-marshmallow==0.9.0
 Flask-Principal==0.4.0
-flask-ldap3-login==0.9.12
-uWSGI==2.0.13.1
-pbr==1.10.0
-python-dateutil==2.5.3
-mysqlclient==1.3.7
-stevedore==1.17.0
-apscheduler==3.2.0
+Flask-RESTful==0.3.7
+Flask-SQLAlchemy==2.3
+humanize==0.5.1
+marshmallow-sqlalchemy==0.15.0
+mysqlclient==1.3.14
+pbr==5.1.1
 prettytable==0.7.2
+python-dateutil==2.7.5
+stevedore==1.30.0
+uWSGI==2.0.17.1
